pyproject.toml
[tool.poetry]
name = "robusta_chaos"
version = "0.0.1"
description = ""
authors = ["Natan Yellin <aantn@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
robusta-cli = "^0.8.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
